+++ src/noggit/MapTile.cpp
@@ -662,10 +662,10 @@ void MapTile::saveTile(World* world)
   // MTEX data
   for (auto const& texture : lTextures)
   {
-    lADTFile.Insert(lCurrentPosition, texture.first.size() + 1, texture.first.c_str());
+    lADTFile.Insert(lCurrentPosition, static_cast<unsigned long>(texture.first.size() + 1), texture.first.c_str());
 
-    lCurrentPosition += texture.first.size() + 1;
-    lADTFile.GetPointer<sChunkHeader>(lMTEX_Position)->mSize += texture.first.size() + 1;
+    lCurrentPosition += static_cast<int>(texture.first.size() + 1);
+    lADTFile.GetPointer<sChunkHeader>(lMTEX_Position)->mSize += static_cast<int>(texture.first.size() + 1);
     LogDebug << "Added texture \"" << texture.first << "\"." << std::endl;
   }
 
@@ -681,15 +681,15 @@ void MapTile::saveTile(World* world)
   for (auto it = lModels.begin(); it != lModels.end(); ++it)
   {
     it->second.filenamePosition = lADTFile.GetPointer<sChunkHeader>(lMMDX_Position)->mSize;
-    lADTFile.Insert(lCurrentPosition, it->first.size() + 1, misc::normalize_adt_filename(it->first).c_str());
-    lCurrentPosition += it->first.size() + 1;
-    lADTFile.GetPointer<sChunkHeader>(lMMDX_Position)->mSize += it->first.size() + 1;
+    lADTFile.Insert(lCurrentPosition, static_cast<unsigned long>(it->first.size() + 1), misc::normalize_adt_filename(it->first).c_str());
+    lCurrentPosition += static_cast<int>(it->first.size() + 1);
+    lADTFile.GetPointer<sChunkHeader>(lMMDX_Position)->mSize += static_cast<int>(it->first.size() + 1);
     LogDebug << "Added model \"" << it->first << "\"." << std::endl;
   }
 
   // MMID
   // M2 model names
-  int lMMID_Size = 4 * lModels.size();
+  int lMMID_Size = static_cast<int>(4 * lModels.size());
   lADTFile.Extend(8 + lMMID_Size);
   SetChunkHeader(lADTFile, lCurrentPosition, 'MMID', lMMID_Size);
   lADTFile.GetPointer<MHDR>(lMHDR_Position + 8)->mmid = lCurrentPosition - 0x14;
@@ -718,14 +718,14 @@ void MapTile::saveTile(World* world)
   for (auto& object : lObjects)
   {
     object.second.filenamePosition = lADTFile.GetPointer<sChunkHeader>(lMWMO_Position)->mSize;
-    lADTFile.Insert(lCurrentPosition, object.first.size() + 1, misc::normalize_adt_filename(object.first).c_str());
-    lCurrentPosition += object.first.size() + 1;
-    lADTFile.GetPointer<sChunkHeader>(lMWMO_Position)->mSize += object.first.size() + 1;
+    lADTFile.Insert(lCurrentPosition, static_cast<unsigned long>(object.first.size() + 1), misc::normalize_adt_filename(object.first).c_str());
+    lCurrentPosition += static_cast<int>(object.first.size() + 1);
+    lADTFile.GetPointer<sChunkHeader>(lMWMO_Position)->mSize += static_cast<int>(object.first.size() + 1);
     LogDebug << "Added object \"" << object.first << "\"." << std::endl;
   }
 
   // MWID
-  int lMWID_Size = 4 * lObjects.size();
+  int lMWID_Size = static_cast<int>(4 * lObjects.size());
   lADTFile.Extend(8 + lMWID_Size);
   SetChunkHeader(lADTFile, lCurrentPosition, 'MWID', lMWID_Size);
   lADTFile.GetPointer<MHDR>(lMHDR_Position + 8)->mwid = lCurrentPosition - 0x14;
@@ -740,7 +740,7 @@ void MapTile::saveTile(World* world)
   lCurrentPosition += 8 + lMWID_Size;
 
   // MDDF
-  int lMDDF_Size = 0x24 * lModelInstances.size();
+  int lMDDF_Size = static_cast<int>(0x24 * lModelInstances.size());
   lADTFile.Extend(8 + lMDDF_Size);
   SetChunkHeader(lADTFile, lCurrentPosition, 'MDDF', lMDDF_Size);
   lADTFile.GetPointer<MHDR>(lMHDR_Position + 8)->mddf = lCurrentPosition - 0x14;
@@ -784,7 +784,7 @@ void MapTile::saveTile(World* world)
   LogDebug << "Added " << lID << " doodads to MDDF" << std::endl;
 
   // MODF
-  int lMODF_Size = 0x40 * lObjectInstances.size();
+  int lMODF_Size = static_cast<int>(0x40 * lObjectInstances.size());
   lADTFile.Extend(8 + lMODF_Size);
   SetChunkHeader(lADTFile, lCurrentPosition, 'MODF', lMODF_Size);
   lADTFile.GetPointer<MHDR>(lMHDR_Position + 8)->modf = lCurrentPosition - 0x14;
@@ -847,8 +847,8 @@ void MapTile::saveTile(World* world)
   if (mFlags & 1)
   {
     size_t chunkSize = sizeof(int16_t) * 9 * 2;
-    lADTFile.Extend(8 + chunkSize);
-    SetChunkHeader(lADTFile, lCurrentPosition, 'MFBO', chunkSize);
+    lADTFile.Extend(static_cast<long>(8 + chunkSize));
+    SetChunkHeader(lADTFile, lCurrentPosition, 'MFBO', static_cast<int>(chunkSize));
     lADTFile.GetPointer<MHDR>(lMHDR_Position + 8)->mfbo = lCurrentPosition - 0x14;
 
     int16_t *lMFBO_Data = lADTFile.GetPointer<int16_t>(lCurrentPosition + 8);
@@ -861,7 +861,7 @@ void MapTile::saveTile(World* world)
     for (int i = 0; i < 9; ++i)
       lMFBO_Data[lID++] = (int16_t)mMinimumValues[i].y;
 
-    lCurrentPosition += 8 + chunkSize;
+    lCurrentPosition += static_cast<int>(8 + chunkSize);
   }
 
   //! \todo Do not do bullshit here in MTFX.
@@ -885,7 +885,7 @@ void MapTile::saveTile(World* world)
   }
 #endif
 
-  lADTFile.Extend(lCurrentPosition - lADTFile.data.size()); // cleaning unused nulls at the end of file
+  lADTFile.Extend(static_cast<long>(lCurrentPosition - lADTFile.data.size())); // cleaning unused nulls at the end of file
 
 
   {
